Moved data class Position into its own file from VariantLoadingUtils
